refactor: break apart over saturated EventListener.java
